+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Next version
 - Added tests showing that ``.descendants().update(...)`` doesn't work, but
   ``.filter(pk__in=....descendants()).update(...)`` does.
 - Added Python 3.13 to the testsuite.
+- Converted the tests to use pytest.
+- Added a ``tree_info`` template tag.
 
 
 0.19 (2024-04-25)